Move socket.io resource management to adapter
Managing the expectations of an ember-data adapter and the event stream
nature of socket.io We can only allow one response in flight at a time.
If more then on go out then the reply of the first promise will also be
captured by the second promise. Meaning that saving model A and model B
at the same time will resolve both A and B to A.

Previously we forced the consumer to manage this using a reduce pattern
to chain the promises. This introduces knowledge coupling to the
consumer and also means that if there were multiple consumers they would
have to coordinate this caveat.

Instead this commit has the chaining logic baked into the adapter and
the call sites revert to how ember-data initially intended.